Back Button & Refactored code in Review page
Made back button functionality when the user is on the review page. Updated / to use params after the initial repo search, so when the user goes back, the field will already be populated.
Refactored code in Review page to be a bit more organized. Made a few components & updates to them, mainly the percentage bar & Modal. Also moved percentage functions to lib file